requirements.txt
addict==2.4.0
asgiref==3.6.0
attrs==23.2.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==3.0.0
Django==3.2.19
django-adminlte3==0.1.6
django-allauth==0.54.0
django-auth==0.1.9
django-ckeditor==6.3.2
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-seed==0.3.1
django-templated-mail==1.1.1
django-url-filter==0.3.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-autodocs==0.4.4
drf-yasg==1.21.7
enum-compat==0.0.3
Faker==18.9.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.34.0
requests==2.29.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
toposort==1.10
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.15
